Use sockets instead of Boost message queues for IPC
This is a lot more stable on the UI side.

+++ engine/src/modules/processing_graph/topology/anthem_graph_node.h
@@ -52,5 +52,5 @@ public:
 
   void initializePorts();
 
-  void setParameter(int index, float value);
+  void setParameter(uint64_t index, float value);
 };
